+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - New #19: Add `sendAudio` method.
 - New #20: Add `sendDocument` method.
 - New #21: Add `sendVideo` method.
+- New #22: Add `getStarTransactions` method and `StarTransactions`, `StarTransaction`, `TransactionPartner`, 
+  `RevenueWithdrawalState` types. 
 
 ## 0.1.0 June 10, 2024